Update the USU join link
The old USU link now goes to a 404. I think they changed it when general
membership became free. I'm not sure the page I found is quite equivalent to
whatever we linked to before, but just from navigating their site it seems to
be where they want you to go to join.